+++ history.md
@@ -56,6 +56,7 @@
 - 4.0.0-preview2
   - *breaking:* Removed `TypeGenerator.GeneratedType[Async]` property. It is substituted by the `GetGeneratedType[Async]()` method
   - *breaking:* Assembly cache directory can be set from `appname.runtimeconfig.json`
+  . *introduced:* Proxy embedding
   - *introduced:* Cancellable `TypeGenerator`s
   - *introduced:* Pointer type support
   - *introduced:* `ref readonly` support